+++ R/haven-sas.R
@@ -96,6 +96,11 @@ write_sas <- function(data, path) {
 #' The SAS transport format is a open format, as is required for submission
 #' of the data to the FDA.
 #'
+#' Note that character limits are expressed in bytes. The number of bytes
+#' will often be the same as the number of characters , but strings with
+#' multibyte character sequences will count some symbols as more than one
+#' character. For example, the string "café" is 5 bytes long in UTF-8.
+#' 
 #' @inheritParams read_spss
 #' @return A tibble, data frame variant with nice defaults.
 #'
@@ -104,7 +109,7 @@ write_sas <- function(data, path) {
 #'
 #'   If a dataset label is defined, it will be stored in the "label" attribute
 #'   of the tibble.
-#'
+#'   
 #'   `write_xpt()` returns the input `data` invisibly.
 #' @export
 #' @examples
@@ -183,11 +188,11 @@ validate_sas <- function(data) {
 
 validate_xpt_name <- function(name, version, call = caller_env()) {
   if (version == 5) {
-    if (nchar(name) > 8) {
+    if (nchar(name, type = "bytes") > 8) {
       cli_abort("{.arg name} must be 8 characters or fewer.", call = call)
     }
   } else {
-    if (nchar(name) > 32) {
+    if (nchar(name, type = "bytes") > 32) {
       cli_abort("{.arg name} must be 32 characters or fewer.", call = call)
     }
   }
@@ -198,7 +203,7 @@ validate_xpt_label <- function(label, call = caller_env()) {
   if (!is.null(label)) {
     stopifnot(is.character(label), length(label) == 1)
 
-    if (nchar(label) > 40) {
+    if (nchar(label, type = "bytes") > 40) {
       cli_abort("{.arg label} must be 40 characters or fewer.", call = call)
     }
   }